requirements.txt
fastapi
uvicorn[standard]
langchain>=0.3.0,<0.4.0
langchain-anthropic>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-chroma>=0.1.0,<0.3.0
langchain-huggingface>=0.1.0,<0.2.0
chromadb>=0.5.0,<0.7.0
sentence-transformers>=3.0.0,<4.0.0
anthropic>=0.40.0,<0.60.0
pydantic-settings
httpx
python-multipart

# Testing (dev only — not needed in production image)
# pytest
# pytest-asyncio
